refactor(auth): Extract session schema and unify base URL
