+++ src/domains/auth/schemas/login-schema.ts
@@ -14,6 +14,10 @@
  */
 import { z } from 'zod'
 
+// Re-exported so consumers importing from `@domains/auth/schemas/login-schema`
+// keep access to the session response schema.
+export * from '@domains/auth/schemas/session-schema'
+
 /**
  * Validates email/password login request bodies.
  *
@@ -75,54 +79,6 @@ export const registerSchema = z.object({
     /** Display name; required, at least one character. */
     name: z.string().min(1),
   }),
-})
-
-/**
- * Validates session lookup responses from Better Auth `getSession`.
- *
- * @remarks
- * **Fields**:
- * - `user` — nullable object:
- *   - `id` — `string`, user identifier
- *   - `email` — `string`, user email
- *   - `name` — `string`, display name
- * - `session` — nullable object:
- *   - `id` — `string`, session identifier
- *   - `userId` — `string`, owning user id
- *   - `expiresAt` — `Date`, coerced from ISO string via `z.coerce.date()`
- *
- * Both `user` and `session` are `null` when the request is unauthenticated.
- *
- * @see {@link sessionService.getSession} — produces data matching this shape
- * @see {@link resolveAuthActor} — extracts `user` and `session` for Astro locals
- *
- * @example
- * ```typescript
- * const sessionData = await sessionService.getSession(request.headers)
- * const validated = sessionResponseSchema.parse(sessionData)
- * ```
- */
-export const sessionResponseSchema = z.object({
-  user: z
-    .object({
-      /** Better Auth user identifier. */
-      id: z.string(),
-      /** User email address. */
-      email: z.string(),
-      /** User display name. */
-      name: z.string(),
-    })
-    .nullable(),
-  session: z
-    .object({
-      /** Better Auth session identifier. */
-      id: z.string(),
-      /** ID of the user this session belongs to. */
-      userId: z.string(),
-      /** Session expiration timestamp (coerced to Date). */
-      expiresAt: z.coerce.date(),
-    })
-    .nullable(),
 })
 
 /**
