+++ src/lib/auth/server.ts
@@ -37,7 +37,7 @@ import {
  */
 export const auth = betterAuth({
   secret: env.BETTER_AUTH_SECRET,
-  baseURL: env.BETTER_AUTH_URL,
+  baseURL: env.APP_BASE_URL,
   emailAndPassword: {
     enabled: true,
   },
